requirements.txt
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
frozendict==2.4.6
idna==3.10
multitasking==0.0.11
numpy==2.1.0
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scipy==1.15.1
six==1.17.0
soupsieve==2.6
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yfinance==0.2.54
alpaca-py
